+++ test/models/job.ts
@@ -1,19 +1,20 @@
-import { is, a, an, namedSchema, nestedArray } from '../../src/index';
+import { is, a, an, namedSchema, nestedArray, schema } from "../../src/index";
 
+@schema()
 export class Office {
   constructor(args: { name: string; location: string }) {
     this.name = args.name;
     this.location = args.location;
   }
 
-  @is(a.string().required('Office name is required'))
+  @is(a.string().required("Office name is required"))
   name: string;
 
-  @is(a.string().required('Office location is required'))
+  @is(a.string().required("Office location is required"))
   location: string;
 }
 
-@namedSchema('job')
+@namedSchema("job")
 export class Job {
   constructor(args: { jobTitle: string; office: Office[] }) {
     this.jobTitle = args.jobTitle;
@@ -23,11 +24,11 @@ export class Job {
   @is(
     a
       .string()
-      .uppercase('Job title must be upper case')
-      .required('Job title is required and must be upper case')
+      .uppercase("Job title must be upper case")
+      .required("Job title is required and must be upper case"),
   )
   jobTitle: string;
 
-  @nestedArray(() => Office, an.array().min(1, 'Office is required'))
+  @nestedArray(() => Office, an.array().min(1, "Office is required"))
   office: Office[];
 }