+++ test/models/person.ts
@@ -1,7 +1,7 @@
-import { a, is, namedSchema, nested } from '../../src/index';
-import { House } from './house';
+import { a, is, namedSchema, nestedType } from "../../src/index";
+import { House } from "./house";
 
-@namedSchema('person')
+@namedSchema("person")
 export class Person {
   constructor(args: { email: string; age: number; house?: House }) {
     this.email = args.email;
@@ -9,17 +9,12 @@ export class Person {
     this.house = args.house;
   }
 
-  @is(a.string().email('Not a valid email'))
+  @is(a.string().email("Not a valid email"))
   email: string;
 
-  @is(
-    a
-      .number()
-      .lessThan(100)
-      .moreThan(0)
-  )
+  @is(a.number().lessThan(100).moreThan(0))
   age: number;
 
-  @nested()
+  @nestedType(() => House)
   house?: House;
 }